+++ window/doc.go
@@ -1,2 +1,2 @@
-// Package window provides an OpenGL window system for free drawing.
+// Package window provides an OpenGL window system for plotting and free drawing.
 package window